refactor(folders): 도메인 구조에서 계층형 구조로 파일 위치 변경
개요

- 기존 Monolithic에서는 도메인 구조의 폴더 스트럭쳐를 썼다.
- 이제 MSA 구조를 쓰니, Member에 관한 것은 확실하므로, 계층의 기능별로
나눈 계층형 구조 structure로 폴더를 변경하자
