+++ spring-rest-member/src/main/java/com/dife/member/service/MemberService.java
@@ -1,5 +1,6 @@
-package com.dife.member.member;
+package com.dife.member.service;
 
+import com.dife.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
