+++ spring-rest-member/src/main/java/dto/MemberDto.java
@@ -1,5 +1,6 @@
-package com.dife.member.model;
+package dto;
 
+import com.dife.member.model.MBTI_category;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
